+++ src/Applet/Test/Upload/appletTestRunCommand.ts
@@ -1,8 +1,7 @@
 import chalk from 'chalk';
 import prompts from 'prompts';
-import { createOrganizationRestApi } from '../../../helper';
+import { createOrganizationRestApiFromUid } from '../../../helper';
 import {
-	getOrganization,
 	getOrganizationUidOrDefaultOrSelect,
 	NO_DEFAULT_ORGANIZATION_OPTION,
 	ORGANIZATION_UID_OPTION,
@@ -73,8 +72,7 @@ export const appletTestRun = createCommandDefinition({
 		await validateAppletDirectory(currentDirectory);
 
 		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options, skipConfirmation);
-		const organization = await getOrganization(organizationUid);
-		const restApi = await createOrganizationRestApi(organization);
+		const restApi = await createOrganizationRestApiFromUid(organizationUid);
 		const deviceUid = await getDeviceUid(restApi, options);
 		const device = await restApi.device.get(deviceUid);