+++ src/Device/Connect/connectCommand.ts
@@ -1,12 +1,11 @@
 import { DEVICE_UID_OPTION, getDeviceUid } from '../deviceFacade';
 import {
-	getOrganization,
 	getOrganizationUidOrDefaultOrSelect,
 	NO_DEFAULT_ORGANIZATION_OPTION,
 	ORGANIZATION_UID_OPTION,
 } from '../../Organization/organizationFacade';
 import { APPLET_UID_OPTION, getAppletUid, getAppletVersion } from '../../Applet/appletFacade';
-import { createOrganizationRestApi, getApiUrl } from '../../helper';
+import { createOrganizationRestApiFromUid, getApiUrl } from '../../helper';
 import { CommandLineOptions, createCommandDefinition } from '../../Command/commandDefinition';
 import { createDevelopmentWithOptions } from '../../Development/developmentFactory';
 import wait from '../../Timer/wait';
@@ -106,11 +105,10 @@ export const connect = createCommandDefinition({
 	run: async function (options: CommandLineOptions<typeof OPTION_LIST>) {
 		const currentDirectory = process.cwd();
 		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
-		const organization = await getOrganization(organizationUid);
-		const restApi = await createOrganizationRestApi(organization);
+		const restApi = await createOrganizationRestApiFromUid(organizationUid);
 		const config = await loadConfig();
 		const dev = createDevelopmentWithOptions({
-			organizationUid: organization.uid,
+			organizationUid,
 			url: getApiUrl(config),
 			accessToken: config.accessToken,
 		});