+++ src/Runner/Upload/runnerUploadCommand.ts
@@ -3,9 +3,8 @@ import { isDeepStrictEqual } from 'util';
 import debug from 'debug';
 import { log } from '@signageos/sdk/dist/Console/log';
 import { CommandLineOptions, createCommandDefinition } from '../../Command/commandDefinition';
-import { createOrganizationRestApi } from '../../helper';
+import { createOrganizationRestApiFromUid } from '../../helper';
 import {
-	getOrganization,
 	getOrganizationUidOrDefaultOrSelect,
 	NO_DEFAULT_ORGANIZATION_OPTION,
 	ORGANIZATION_UID_OPTION,
@@ -54,8 +53,7 @@ export const runnerUpload = createCommandDefinition({
 		const currentDirectory = process.cwd();
 		const skipPrompts = options.yes as boolean;
 		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options, skipPrompts);
-		const organization = await getOrganization(organizationUid);
-		const restApi = await createOrganizationRestApi(organization);
+		const restApi = await createOrganizationRestApiFromUid(organizationUid);
 
 		const config = await getSosConfig(currentDirectory);
 		const schema = await loadSchemas(currentDirectory);